chore: some renaming
