+++ pkg/cache/strategy/redis/options.go
@@ -1,4 +1,4 @@
-package generic_cache_redis
+package redis
 
 const (
 	defaultHost     = "localhost"
